==> 01_setup.ipynb <==
!pip install matplotlib==3.9.1
!pip install matplotlib-inline==0.1.7
!pip install numpy==1.26.0
!pip install openpyxl==3.1.5
!pip install pandas==2.2.2
!pip install pandera==0.19.3
!pip install pyarrow==16.1.0
!pip install scipy==1.14.0
!pip install scikit-learn==1.3.2
!pip install seaborn==0.13.2

pip list
