==> 03_import_data_part_2.ipynb <==
import pandas as pd
import warnings
warnings.filterwarnings("ignore", category=FutureWarning)
warnings.filterwarnings("ignore", category=RuntimeWarning)
warnings.filterwarnings("ignore", category=UserWarning)
pd.options.mode.chained_assignment = None  # default='warn'
import os

def Concatenate_new(base_directory,mode):
    # List to store dataframes
    dataframes = []
    # Traverse the directory
    for root, dirs, files in os.walk(base_directory):
        for file in files:
            if file == mode:
                file_path = os.path.join(root, file)
                # Read the csv file
                df = pd.read_csv(file_path, converters={'CustomerID': str, 'CustomerRefID': str, 'RequestID': str,'RegistrationID':str})
                if 'RatingDate' in df.columns: df['RatingDate'] = pd.to_datetime(df['RatingDate'])
                if 'RequestDate' in df.columns: df['RequestDate'] = pd.to_datetime(df['RequestDate'])
                   
                # Append the dataframe to the list
                dataframes.append(df)

    # Concatenate all dataframes into one
    final_dataframe = pd.concat(dataframes, ignore_index=True)
    return final_dataframe


def Concatenate_old(base_directory,mode):
    dataframes = []
    for root, dirs, files in os.walk(base_directory):
        # Skip the 'No' folder
        if '26.Dec22' in root.split(os.sep):
            continue
        for file in files:
            if file == mode:
                file_path = os.path.join(root, file)
                # Read the csv file
                df = pd.read_csv(file_path, converters={'CustomerID': str, 'CustomerRefID': str, 'RequestID': str,'RegistrationID':str})
                if 'RatingDate' in df.columns: df['RatingDate'] = pd.to_datetime(df['RatingDate'])
                if 'RequestDate' in df.columns: df['RequestDate'] = pd.to_datetime(df['RequestDate'],format="%d/%m/%Y")
                # Append the dataframe to the list
                dataframes.append(df)
    
    final_dataframe = pd.concat(dataframes, ignore_index=True)
    return final_dataframe

def run1(a,mode):
    a.reset_index(drop=True,inplace=True)
    # a.to_csv(f"../data/processed/01_master_data/{mode}_Dec22_Mar24.csv",encoding='utf-8-sig')
    a.to_parquet(f"../data/processed/01_master_data/{mode}_Dec22_Mar24.parquet")

def run(a,mode):
    a.reset_index(drop=True,inplace=True)
    # a.to_csv(f"../data/processed/01_master_data/{mode}_old_Nov22.csv",encoding='utf-8-sig')
    a.to_parquet(f"../data/processed/01_master_data/{mode}_old_Nov22.parquet")
   




#ของใหม่
direc="../data/raw/Dec22_Mar24"
a=Concatenate_new(direc,'ScoreResult_Composite.csv')
b=Concatenate_new(direc,'ScoreResult_Business.csv')
d=Concatenate_new(direc,'ScoreResult_Financial.csv')
e=Concatenate_new(direc,'ScoreResult_BusinessSMEs.csv')
f=Concatenate_new(direc,'ScoreResult_Industry.csv')
run1(a,'ScoreResult_Composite')
run1(b,'ScoreResult_Business')
run1(d,'ScoreResult_Financial')
run1(e,'ScoreResult_BusinessSMEs')
run1(f,'ScoreResult_Industry')


#ของเก่า
direc2="../data/raw/Score Result ก่อนปี 2566"
a=Concatenate_old(direc2,'ScoreResult_Composite.csv')
b=Concatenate_old(direc2,'ScoreResult_Business.csv')
d=Concatenate_old(direc2,'ScoreResult_Financial.csv')
e=Concatenate_old(direc2,'ScoreResult_Express.csv')
f=Concatenate_old(direc2,'ScoreResult_Industry.csv')
run(a,'ScoreResult_Composite')
run(b,'ScoreResult_Business')
run(d,'ScoreResult_Financial')
run(e,'ScoreResult_Express')
run(f,'ScoreResult_Industry')

print('finish')
